Extract the register logic to a service
